Add Upload and Category Object on Responses

+++ src/services/DeleteTransactionService.ts
@@ -2,11 +2,20 @@ import { getCustomRepository } from 'typeorm';
 
 import TransactionsRepository from '../repositories/TransactionsRepository';
 
+import AppError from '../errors/AppError';
+
 class DeleteTransactionService {
   public async execute(id: string): Promise<void> {
     const transactionsRepository = getCustomRepository(TransactionsRepository);
 
-    await transactionsRepository.delete(id);
+    const transaction = await transactionsRepository.findOne(id);
+
+    if (transaction) {
+      await transactionsRepository.delete(id);
+    }
+    if (!transaction) {
+      throw new AppError('Transaction does not exist.', 400);
+    }
   }
 }
